+++ src/main/scala/tscfg/example/ScalaExample0Cfg.scala
@@ -3,7 +3,6 @@ package tscfg.example
 case class ScalaExample0Cfg(
   service : ScalaExample0Cfg.Service
 )
-
 object ScalaExample0Cfg {
   case class Service(
     debug    : scala.Boolean,
@@ -11,20 +10,20 @@ object ScalaExample0Cfg {
     poolSize : scala.Int,
     url      : java.lang.String
   )
-
   object Service {
-    def apply(c: com.typesafe.config.Config): Service = {
-      Service(
-        c.getBoolean("debug"),
-        c.getDouble("factor"),
-        c.getInt("poolSize"),
-        c.getString("url")
+    def apply(c: com.typesafe.config.Config): ScalaExample0Cfg.Service = {
+      ScalaExample0Cfg.Service(
+        debug    = if(c.hasPathOrNull("debug")) c.getBoolean("debug") else true,
+        factor   = if(c.hasPathOrNull("factor")) c.getDouble("factor") else 0.75,
+        poolSize = if(c.hasPathOrNull("poolSize")) c.getInt("poolSize") else 32,
+        url      = if(c.hasPathOrNull("url")) c.getString("url") else "http://example.net/rest"
       )
     }
   }
+        
   def apply(c: com.typesafe.config.Config): ScalaExample0Cfg = {
     ScalaExample0Cfg(
-      Service(c.getConfig("service"))
+      service = ScalaExample0Cfg.Service(c.getConfig("service"))
     )
   }
 }