+++ src/main/scala/tscfg/generators/JavaGenMain.scala
@@ -26,8 +26,8 @@ object JavaGenMain {
     val genOpts = GenOpts("tscfg.example", className, j7 = false)
     val gen = new JavaGen(genOpts)
     val res = gen.generate(objectType)
-    println(s"classNames = ${res.classNames}")
-    println(s"fieldNames = ${res.fieldNames}")
+    println(s"classNames: ${res.classNames}")
+    println(s"fields    : ${res.fields}")
 
     import _root_.java.io._
     val destFilename  = s"src/main/java/tscfg/example/JCfg.java"