+++ changelog.md
@@ -1,3 +1,8 @@
+2016-12-14 - 0.7.1
+- fixed: spec with literal values were not generating "primitive" types 
+  in java. There was a missing `&& a.default.isEmpty` in:
+  `val typ = if (a.optional && a.default.isEmpty) toObjectType(memberType) else memberType`
+
 2016-12-12 - 0.7.0
 - capture model build warnings